remove extra work in inner function sweep
All exploration and validation are already handled in the assume function of the Solver class (the pattern in the previous code was suspiciously similar). Because each necessary result builds upon the previous results in the sweep (because they are applied immediately), there is no exploration or validation needed for all of the results as a whole.

Also, the inner function sweep now always returns a boolean, which it should. This didn't affect the solve function because an object (i.e. a memory address) is truthy, so when sweepResult.valid wasn't the value false itself, it was effectively true.