added ability to re-render board of same size
The old code actually had a bug when render was called and there was already a rendered board of a different size, which is part of what inspired this. Additionally, if the new board is the same size as the already rendered board, then we can simply update all the cell images. Being able to compare the board sizes requires that the 'grid' DOM object "knows" its own dimensions, so to speak.