sweep recursively over necessary results' horizons
In subsequent recursive calls of the inner function sweep, we only need to sweep over the unset groups in the union of the horizons of the necessary results for the current call. We shouldn't be accumulating cells in the set called 'horizon' over all recursive calls. The function sweep of the Solver class also doesn't need to return any horizon, as there evidently isn't any use for it.

The 'mergeInto' function was also a bit inefficient, as the Set.prototype.add function is already idempotent (i.e. prevents duplicate entries, in this case) without the conditional check that the element to be added is not already in the set (the 'add' function handles this itself). As 'mergeInto' was only referenced once and was quite small, it's been replaced there with the inline equivalent of what would be the more efficient version of it (making the code more readable).